+++ app/src/main/AndroidManifest.xml
@@ -24,6 +24,15 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".login.Auth_back_Activity"
+            android:exported="false" />
+        <activity
+            android:name=".login.SignUpActivity"
+            android:exported="true" />
+        <activity
+            android:name=".login.SignInActivity"
+            android:exported="true" />
+        <activity
             android:name=".LocationTestActivity"
             android:exported="true" />
         <activity
