+++ app/src/main/java/com/example/visit_jeju_app/login/SignUpActivity.kt
@@ -1,0 +1,164 @@
+package com.example.visit_jeju_app.login
+
+import android.graphics.Paint
+import android.os.Bundle
+import android.util.Log
+import android.view.View
+import android.widget.TextView
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import com.example.visit_jeju_app.MyApplication
+import com.example.visit_jeju_app.R
+import com.example.visit_jeju_app.databinding.ActivitySignUpBinding
+import com.example.visit_jeju_app.login.model.User
+import com.example.visit_jeju_app.retrofit.addUserToMysql
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+
+class SignUpActivity : AppCompatActivity() {
+    private lateinit var auth: FirebaseAuth
+    lateinit var binding :ActivitySignUpBinding
+
+    // 이용약관 및 개인정보 취급방침 동의 프래그먼트
+    private val Fragment_1 = 1
+    private val Fragment_2 = 2
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivitySignUpBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        auth= Firebase.auth
+
+        binding.signBtn.setOnClickListener {
+            //이메일,비밀번호 회원가입
+            val username = binding.authUsernameEditView.text.toString()
+            val email = binding.authEmailEditView.text.toString()
+            val password = binding.authPasswordEditView.text.toString()
+
+            MyApplication.auth.createUserWithEmailAndPassword(email, password)
+                .addOnCompleteListener(this) { task ->
+                    if (task.isSuccessful) {
+                        // Firebase 회원가입 성공
+                        val userId = MyApplication.auth.currentUser?.uid ?: ""
+
+                        // 이메일 인증 보내기
+                        MyApplication.auth.currentUser?.sendEmailVerification()
+                            ?.addOnCompleteListener { sendTask ->
+                                if (sendTask.isSuccessful) {
+                                    addUserToDatabase(username, email, auth.currentUser?.uid!!)
+                                    // 이메일 인증 성공, MySQL에 사용자 정보 저장
+                                    addUserToMysql(username, email, userId)
+                                    Log.d("lsy","1차 확인")
+                                    Toast.makeText(
+                                        baseContext,
+                                        "회원가입에서 성공, 전송된 메일을 확인해 주세요",
+                                        Toast.LENGTH_SHORT
+                                    ).show()
+                                    changeVisibility("logout")
+                                } else {
+                                    // 이메일 인증 실패
+                                    Toast.makeText(baseContext, "메일 발송 실패", Toast.LENGTH_SHORT)
+                                        .show()
+                                    changeVisibility("logout")
+                                    Log.d("lsy","1차 확인 실패")
+                                }
+                            }
+                    } else {
+                        // Firebase 회원가입 실패
+                        Toast.makeText(baseContext, "회원가입 실패", Toast.LENGTH_SHORT).show()
+                        changeVisibility("logout")
+                        Log.d("lsy","1차 확인실패 2")
+                    }
+                }
+        }
+
+        // 이용약관 및 개인정보처리방침 동의
+        //체크박스 클릭 시 이벤트 핸들러
+        binding.checkboxUserAgreement.setOnCheckedChangeListener { buttonView, isChecked ->
+            Toast.makeText(this@SignUpActivity, "이용약관 동의", Toast.LENGTH_SHORT).show()
+        }
+        binding.checkboxPrivacyPolicy.setOnCheckedChangeListener { buttonView, isChecked ->
+            Toast.makeText(this@SignUpActivity, "개인정보처리방침 동의", Toast.LENGTH_SHORT).show()
+        }
+        // 첫 번째 TextView에 밑줄 추가
+        val btnUserAgreement = findViewById<TextView>(R.id.btnUserAgreement)
+        btnUserAgreement.paintFlags = btnUserAgreement.paintFlags or Paint.UNDERLINE_TEXT_FLAG
+
+        // 두 번째 TextView에 밑줄 추가
+        val btnPrivacyPolicy = findViewById<TextView>(R.id.btnPrivacyPolicy)
+        btnPrivacyPolicy.paintFlags = btnPrivacyPolicy.paintFlags or Paint.UNDERLINE_TEXT_FLAG
+
+
+
+
+    } //onCreate
+
+    //이용약관, 개인정보취급방침 프래그먼트 호출
+/*    fun openFragment(view: View) {
+        val fragment = when(view.id) {
+            R.id.btnPrivacyPolicy -> PrivacyPolicyFragment() // PrivacyPolicyFragment()는 프래그먼트 클래스 이름입니다.
+            R.id.btnUserAgreement -> UserAgreementFragment() // UserAgreementFragment()는 프래그먼트 클래스 이름입니다.
+            else -> return // 다른 버튼이면 함수를 종료합니다.
+        }
+
+        val fragmentManager: FragmentManager = supportFragmentManager
+        val fragmentTransaction: FragmentTransaction = fragmentManager.beginTransaction()
+        fragmentTransaction.replace(R.id.fragment_container, fragment)
+        fragmentTransaction.addToBackStack(null)
+        fragmentTransaction.commit()
+    }*/
+
+    fun openFragment(view: View) {
+        val fragment = when (view.id) {
+            R.id.btnPrivacyPolicy -> PrivacyPolicyFragment()
+            R.id.btnUserAgreement -> UserAgreementFragment()
+            else -> return
+        }
+
+        fragment.show(supportFragmentManager, "dialog_fragment_tag")
+    }
+
+    private fun saveUser(){
+        val data = mapOf(
+            "email" to binding.authEmailEditView.text.toString(),
+            "password" to binding.authPasswordEditView.text.toString(),
+            "username" to binding.authUsernameEditView.text.toString(),
+        )
+        MyApplication.db.collection("user")
+            .add(data)
+            .addOnFailureListener{
+                Log.d("kkang", "data save error", it)
+            }
+    }
+
+    private fun addUserToDatabase(name: String, email: String, uId: String){
+        MyApplication.rdb.child("user").child(uId).setValue(User(name, email, uId))
+
+    }
+
+    //매개변수를 모드라는 변수명,문자열 타입.
+    fun changeVisibility(mode: String){
+        if(mode === "signin"){
+            binding.run {
+                //authMainTextView.text = "아직 회원이 아니라면 회원가입 해 주세요."
+                //logoutBtn.visibility = View.GONE
+                //goSignInBtn.visibility = View.GONE
+                //authNotEmail.visibility= View.GONE
+                authUsernameEditView.visibility = View.VISIBLE
+                authEmailEditView.visibility = View.VISIBLE
+                authPasswordEditView.visibility = View.VISIBLE
+                authUsernameEditView.visibility= View.VISIBLE
+                signBtn.visibility = View.VISIBLE
+
+                //hostSignBtn.visibility= View.GONE
+                //loginBtn.visibility = View.GONE
+                //googleSignBtn.visibility= View.GONE
+            }
+        }
+    }
+}
+
