+++ app/src/main/res/values/themes.xml
@@ -42,6 +42,20 @@
     <style name="buttonStyle">
         <item name="android:textColor">@color/white</item>
         <item name="android:textSize">16sp</item>
+        <item name="android:background">@drawable/btn_radius_point</item>
+        <item name="android:padding">12dp</item>
+    </style>
+    <style name="buttonStyle2">
+        <item name="android:textColor">@color/white</item>
+        <item name="android:textSize">16sp</item>
+        <item name="android:background">@drawable/btn_radius_base</item>
+        <item name="android:padding">12dp</item>
+    </style>
+
+    <style name="signTitleStyle">
+        <item name="android:textColor">@color/jejutourPoint</item>
+        <item name="android:textSize">22sp</item>
+        <item name="android:textStyle">bold</item>
     </style>
